+++ src/com/wizsharing/entity/Datagrid.java
@@ -15,13 +15,13 @@ public class Datagrid<T> implements Serializable{
 	 */
 	private static final long serialVersionUID = 1L;
 	private List<T>  rows= Collections.emptyList();
-	private Integer total=0;
+	private Long total=0L;
 	
 	public Datagrid(){
 		
 	}
 	
-	public Datagrid(Integer total, List<T> rows){
+	public Datagrid(Long total, List<T> rows){
 		this.rows = rows;
 		this.total = total;
 	}
@@ -29,16 +29,17 @@ public class Datagrid<T> implements Serializable{
 	public List<T> getRows() {
 		return rows;
 	}
+	
 	public void setRows(List<T> rows) {
 		this.rows = rows;
 	}
-	public Integer getTotal() {
+
+	public Long getTotal() {
 		return total;
 	}
-	public void setTotal(Integer total) {
+
+	public void setTotal(Long total) {
 		this.total = total;
 	}
-
-	
 	
 }
